requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.1.2
djangorestframework==3.14.0
drf-yasg==1.21.4
httpie==3.2.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-status==1.0.1
pytz==2022.5
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.0
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.3
swagger-spec-validator==3.0.2
uritemplate==4.1.1
urllib3==1.26.12
